chore: move typings around, remove unused imports, improve type safety
