+++ studio/components/interfaces/Home/ProjectUsage.tsx
@@ -10,11 +10,7 @@ import {
 } from '@supabase/ui'
 import ChartHandler from 'components/to-be-cleaned/Charts/ChartHandler'
 import Panel from 'components/ui/Panel'
-import Table from 'components/to-be-cleaned/Table'
-import { USAGE_COLORS } from 'components/ui/Charts/Charts.constants'
-import StackedAreaChart from 'components/ui/Charts/StackedAreaChart'
 import dayjs from 'dayjs'
-import { useFlag } from 'hooks'
 import { get } from 'lib/common/fetch'
 import { API_URL, DATE_FORMAT, METRICS } from 'lib/constants'
 import Link from 'next/link'
