+++ studio/components/interfaces/Reports/Reports.types.ts
@@ -1,3 +1,17 @@
 export enum Presets {
-    OVERVIEW = "overview"
+  OVERVIEW = 'overview',
+}
+export interface StatusCodesDatum {
+  timestamp: number
+  status_code: number
+  count: number
+}
+export interface PathsDatum {
+  timestamp: number
+  path: string
+  query_params: string
+  count: number
+  sum: number
+  method: string
+  avg_origin_time: number
 }
